requirements.txt
absl-py==0.9.0
AnyQt==0.0.10
apturl==0.5.2
apyori==1.1.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
baycomp==1.0.2
bcrypt==3.1.7
bleach==3.1.1
blinker==1.4
Bottleneck==1.3.2
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dictdiffer==0.8.1
distro==1.4.0
distro-info==0.23
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
ExifRead==2.3.1
fasteners==0.14.1
filelock==3.0.12
Flask==1.0.3
future==0.18.2
gast==0.3.3
gdown==3.12.0
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
html5lib==1.0.1
httpcore==0.10.2
httplib==20.14.0
httpx==0.14.3
idna==2.8
importlib-metadata1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.10.1
joblib==0.16.0
jsonschema==3.2.0
jupyter-client6.1.2
jupyter-console6.0.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.1.0
keras-retinanet==0.5.1
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.2.0
language-selector0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.3.0
mistune==0.8.4
mlxtend==0.17.3
monotonic==1.5
more-itertools4.2.0
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.5
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.3.0.36
openpyxl==3.0.5
openTSNE==0.4.4
opt-einsum==3.3.0
orange-canvas-core0.1.15
orange-widget-base4.7.0
Orange33.26.0
Orange3-Associate1.1.6
pandas==1.1.0
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.0.2
progressbar2==3.51.4
prometheus-client0.7.1
prompt-toolkit2.0.10
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pygame==1.9.6
Pygments==2.3.1
PyGObject==3.36.0
pyheif==0.5.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyqtgraph==0.11.0
pyrsistent==0.15.5
PySocks==1.7.1
python-dateutil2.7.3
python-louvain==0.14
python-utils==2.4.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
serverfiles==0.3.0
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
sniffio==1.1.0
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tornado==5.1.1
tqdm==4.48.2
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib==31.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2
widgetsnbextension==2.0.0
wrapt==1.12.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.3.3
zipp==1.0.0

